# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Horse Racing AI Predictor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /races:
    get:
      operationId: listRaces
      tags: [races]
      summary: List races with optional filtering
      parameters:
        - name: trackId
          in: query
          required: false
          schema:
            type: integer
        - name: date
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [upcoming, completed, all]
      responses:
        "200":
          description: List of races
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Race"
    post:
      operationId: createRace
      tags: [races]
      summary: Create a new race
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRaceBody"
      responses:
        "201":
          description: Created race
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Race"
  /races/{id}:
    get:
      operationId: getRace
      tags: [races]
      summary: Get a race by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Race details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Race"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /races/{id}/entries:
    get:
      operationId: getRaceEntries
      tags: [races]
      summary: Get horses entered in a race
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Race entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RaceEntry"
  /races/{id}/results:
    post:
      operationId: recordRaceResult
      tags: [races]
      summary: Record actual race results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordResultBody"
      responses:
        "200":
          description: Updated race
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Race"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Race:
      type: object
      properties:
        id:
          type: integer
        trackId:
          type: integer
        trackName:
          type: string
        raceNumber:
          type: integer
        raceName:
          type: string
        raceDate:
          type: string
        postTime:
          type: string
        distance:
          type: string
        surface:
          type: string
        purse:
          type: integer
        conditions:
          type: string
        status:
          type: string
          enum: [upcoming, completed, cancelled]
        winnerHorseId:
          type: integer
          nullable: true
        winnerHorseName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, trackId, raceNumber, raceName, raceDate, distance, surface, status, createdAt]
    CreateRaceBody:
      type: object
      properties:
        trackId:
          type: integer
        raceNumber:
          type: integer
        raceName:
          type: string
        raceDate:
          type: string
        postTime:
          type: string
        distance:
          type: string
        surface:
          type: string
          enum: [dirt, turf, synthetic]
        purse:
          type: integer
        conditions:
          type: string
      required: [trackId, raceNumber, raceName, raceDate, distance, surface]
    RecordResultBody:
      type: object
      properties:
        winnerHorseId:
          type: integer
        finishOrder:
          type: array
          items:
            type: integer
      required: [winnerHorseId]
    RaceEntry:
      type: object
      properties:
        id:
          type: integer
        raceId:
          type: integer
        horseId:
          type: integer
        horseName:
          type: string
        postPosition:
          type: integer
        jockey:
          type: string
        trainer:
          type: string
        morningLineOdds:
          type: string
        weight:
          type: integer
        lastRaceDate:
          type: string
          nullable: true
        lastRaceFinish:
          type: integer
          nullable: true
        winPercentage:
          type: number
          nullable: true
        placePercentage:
          type: number
          nullable: true
        totalRaces:
          type: integer
        totalWins:
          type: integer
      required: [id, raceId, horseId, horseName, postPosition, jockey, trainer, totalRaces, totalWins]
...