# lib/api-zod/src/generated/types/recordResultBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse Racing AI Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface RecordResultBody {
  winnerHorseId: number;
  finishOrder?: number[];
}